스토리 버그 수정
